borsh

Module de

Modules

hashes

Module is available if borsh is built with features = ["std"] or features = ["hashbrown"].

Traits

BorshDeserialize

A data-structure that can be de-serialized from binary format by NBOR.

EnumExt

Additional methods offered on enums which is used by [derive(BorshDeserialize)].

Functions

from_reader

Deserializes an object from a reader.

from_slice

Deserializes an object from a slice of bytes.