borsh

Module io

Subset of std::io which is used as part of borsh public API.

When crate is built with std feature disabled (it’s enabled by default), the exported types are custom borsh types which try to mimic behaviour of corresponding standard types usually offering subset of features.

Structs

Error

The error type for I/O operations of the Read, Write, Seek, and associated traits.

Enums

ErrorKind

A list specifying general categories of I/O error.

Traits

Read

The Read trait allows for reading bytes from a source.

Write

A trait for objects which are byte-oriented sinks.

Type Aliases

Result

A specialized Result type for I/O operations.