Trait borsh::de::BorshDeserialize

pub trait BorshDeserialize: Sized {
    fn deserialize(buf: &mut &[u8]) -> Result<Self>;

    fn try_from_slice(v: &[u8]) -> Result<Self> { ... }
}

A data-structure that can be de-serialized from binary format by NBOR.

Required methods

fn deserialize(buf: &mut &[u8]) -> Result<Self>

Deserializes this instance from a given slice of bytes. Updates the buffer to point at the remaining bytes.

Provided methods

fn try_from_slice(v: &[u8]) -> Result<Self>

Deserialize this instance from a slice of bytes.